+++ classes/observers.php
@@ -49,6 +49,7 @@ use local_booking\local\participant\entities\student;
 use local_booking\local\session\entities\booking;
 use core\event\course_module_completion_updated;
 use core\event\course_updated;
+use core\event\user_graded;
 use core\session\exception;
 use mod_assign\event\submission_graded;
 use mod_lesson\event\lesson_ended;
@@ -171,28 +172,6 @@ class observers {
             } else {
                 throw new exception(get_string('errorgradeitemnotfound', 'local_booking'));
             }
-
-            // Since you can't directly redirect from the observer (it runs in the background)
-            Global $SESSION;
-
-            // set the from page to 'mod-assign-view'
-            $SESSION->frompage = 'mod-assign-view';
-
-            // redirect to the checklist grading page for the student if the course has checklists otherwise redirect to the booking view page for the student
-            $haschecklists = subscriber::has_checklists($courseid);
-            if ($haschecklists) {
-                $redirecturl = new \moodle_url('/local/booking/checklist_grading.php', [
-                    'courseid' => $courseid,
-                    'studentid' => $studentid,
-                    'bookingid' => $booking->get_id(),
-                ]);
-            } else {
-                $redirecturl = new \moodle_url('/local/booking/view.php', [
-                    'courseid' => $courseid,
-                ]);
-            }
-            // Set a flag in session to trigger redirect
-            $SESSION->booking_redirect = $redirecturl->out(false);
         }
     }
 }